The perspective actually works?

+++ example/boxes_fs.glsl
@@ -1,11 +1,13 @@
 #version 300 es
 
-precision mediump float;
+precision highp float;
 
+// the varied color passed from the vertex shader
 in vec4 v_color;
 
+// we need to declare an output for the fragment shader
 out vec4 out_color;
 
 void main() {
-	out_color = v_color;
+  out_color = v_color;
 }
